refactor: move loaders to a separate folder
